move normalized structs to models file
